Move child spawning functions to utils/child_process.ts
